Extracted authentication and redirection logic into a reusable useAuthGuard hook.
