+++ frontend/app.json
@@ -2,7 +2,7 @@
   "expo": {
     "name": "0km-app",
     "slug": "0km-app",
-    "entryPoint": "./src/app/index.tsx",
+    "entryPoint": "./src/app/_layout.tsx",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./src/assets/images/logo.png",
